truthlens-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
google-generativeai==0.8.0
requests==2.31.0
newspaper3k==0.2.8
beautifulsoup4==4.12.2
python-whois==0.9.4
pytesseract==0.3.10
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
librosa==0.10.1
soundfile==0.12.1
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
aiosqlite==0.19.0
